# doc/openapi/payment-gateway.yaml
openapi: 3.1.0
info:
  title: Payment Gateway API
  description: |
    Checkout.com Payment Gateway — processes card payments through an acquiring bank
    and stores results in memory. Merchants can submit payments and retrieve
    previously made payments by ID.
  version: 0.6.0
  contact:
    name: Checkout.com

servers:
  - url: http://localhost:8090
    description: Local development

paths:
  /v1/payment:
    post:
      operationId: processPayment
      summary: Process a card payment
      description: |
        Submit a card payment for processing. The gateway validates the request,
        forwards it to the acquiring bank, and returns the result.

        Possible outcomes:
        - **Authorized** (201) — bank approved, payment stored
        - **Declined** (201) — bank rejected, payment stored
        - **Rejected** (400) — validation failed, bank NOT called, payment NOT stored
      tags:
        - Payments
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: |
            Optional UUID to ensure idempotent processing. If a payment with the same
            key has already been processed, the cached response is returned without
            calling the bank again. Recommended for retry-safe integrations.
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessPaymentRequest'
            examples:
              valid_authorized:
                summary: Valid card (will be authorized — odd last digit)
                value:
                  cardNumber: "2222405343248877"
                  expiryMonth: 4
                  expiryYear: 2028
                  currency: "GBP"
                  amount: 100
                  cvv: "123"
              valid_declined:
                summary: Valid card (will be declined — even last digit)
                value:
                  cardNumber: "2222405343248878"
                  expiryMonth: 4
                  expiryYear: 2028
                  currency: "USD"
                  amount: 5000
                  cvv: "456"
              invalid_card:
                summary: Invalid card number (too short)
                value:
                  cardNumber: "123456789"
                  expiryMonth: 12
                  expiryYear: 2030
                  currency: "EUR"
                  amount: 100
                  cvv: "789"
      responses:
        '201':
          description: Payment processed (Authorized or Declined)
          headers:
            Location:
              schema:
                type: string
              description: URI of the created payment resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessPaymentResponse'
              examples:
                authorized:
                  summary: Authorized
                  value:
                    id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    status: "Authorized"
                    cardNumberLastFour: "8877"
                    expiryMonth: 4
                    expiryYear: 2028
                    currency: "GBP"
                    amount: 100
                declined:
                  summary: Declined
                  value:
                    id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                    status: "Declined"
                    cardNumberLastFour: "8878"
                    expiryMonth: 4
                    expiryYear: 2028
                    currency: "USD"
                    amount: 5000
        '400':
          description: Payment rejected (validation failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              example:
                status: "Rejected"
                message: "Validation failed"
                errors:
                  - field: "cardNumber"
                    message: "size must be between 14 and 19"
        '429':
          description: Rate limited
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until the client should retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Rate limit exceeded"
        '502':
          description: Bank communication failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Bank service unavailable"

  /v1/payment/{id}:
    get:
      operationId: getPaymentById
      summary: Retrieve a previously made payment
      description: |
        Retrieve details of a previously made payment using its identifier.
        Helps merchants with reconciliation and reporting needs.
      tags:
        - Payments
      parameters:
        - name: id
          in: path
          required: true
          description: Payment UUID
          schema:
            type: string
            format: uuid
            example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
      responses:
        '200':
          description: Payment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentDetailsResponse'
              example:
                id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                status: "Authorized"
                cardNumberLastFour: "8877"
                expiryMonth: 4
                expiryYear: 2027
                currency: "GBP"
                amount: 100
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Payment not found"
        '429':
          description: Rate limited
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until the client should retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Rate limit exceeded"

components:
  schemas:
    ProcessPaymentRequest:
      type: object
      required:
        - cardNumber
        - expiryMonth
        - expiryYear
        - currency
        - amount
        - cvv
      properties:
        cardNumber:
          type: string
          description: Full card number (14-19 digits, numeric only, must pass Luhn check). Never stored or returned.
          minLength: 14
          maxLength: 19
          pattern: '^\d{14,19}$'
          example: "2222405343248877"
        expiryMonth:
          type: integer
          description: Card expiry month (1-12)
          minimum: 1
          maximum: 12
          example: 4
        expiryYear:
          type: integer
          description: Card expiry year. Combined with expiryMonth, must be in the future.
          minimum: 2024
          example: 2028
        currency:
          type: string
          description: ISO 4217 currency code. Supported values — GBP, USD, EUR
          enum:
            - GBP
            - USD
            - EUR
          example: "GBP"
        amount:
          type: integer
          description: Amount in minor currency units (e.g. $10.50 = 1050). Must be positive.
          minimum: 1
          example: 100
        cvv:
          type: string
          description: Card verification value (3-4 digits, numeric only). Never stored or returned.
          minLength: 3
          maxLength: 4
          pattern: '^\d{3,4}$'
          example: "123"

    ProcessPaymentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Payment ID generated by the gateway
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        status:
          type: string
          description: Payment outcome
          enum:
            - Authorized
            - Declined
            - Rejected
          example: "Authorized"
        cardNumberLastFour:
          type: string
          description: Last 4 digits of the card number (PCI compliant)
          minLength: 4
          maxLength: 4
          pattern: '^\d{4}$'
          example: "8877"
        expiryMonth:
          type: integer
          description: Card expiry month (1-12)
          example: 4
        expiryYear:
          type: integer
          description: Card expiry year
          example: 2028
        currency:
          type: string
          description: ISO 4217 currency code
          example: "GBP"
        amount:
          type: integer
          description: Amount in minor currency units
          example: 100

    PaymentDetailsResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Payment ID
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        status:
          type: string
          description: Payment status (only Authorized or Declined — Rejected payments are not stored)
          enum:
            - Authorized
            - Declined
          example: "Authorized"
        cardNumberLastFour:
          type: string
          description: Last 4 digits of the card number (PCI compliant)
          minLength: 4
          maxLength: 4
          pattern: '^\d{4}$'
          example: "8877"
        expiryMonth:
          type: integer
          description: Card expiry month (1-12)
          example: 4
        expiryYear:
          type: integer
          description: Card expiry year
          example: 2028
        currency:
          type: string
          description: ISO 4217 currency code
          example: "GBP"
        amount:
          type: integer
          description: Amount in minor currency units
          example: 100

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error description
          example: "Payment not found"

    FieldError:
      type: object
      properties:
        field:
          type: string
          description: Name of the field that failed validation
          example: "cardNumber"
        message:
          type: string
          description: Description of the validation failure
          example: "size must be between 14 and 19"

    ValidationErrorResponse:
      type: object
      properties:
        status:
          type: string
          description: Always "Rejected" for validation errors
          enum:
            - Rejected
          example: "Rejected"
        message:
          type: string
          description: Summary of the error
          example: "Validation failed"
        errors:
          type: array
          description: List of field-level validation errors
          items:
            $ref: '#/components/schemas/FieldError'